use vault-owned generator component
