+++ apps/desktop/src/app/tools/generator/credential-generator.component.html
@@ -1,12 +1,7 @@
 <bit-dialog #dialog dialogSize="large" background="alt">
   <span bitDialogTitle>{{ "generator" | i18n }}</span>
   <ng-container bitDialogContent>
-    <ng-container *ngIf="data.type === 'username'">
-      <tools-username-generator (onGenerated)="onCredentialGenerated($event)" />
-    </ng-container>
-    <ng-container *ngIf="data.type === 'password'">
-      <tools-password-generator (onGenerated)="onCredentialGenerated($event)" />
-    </ng-container>
+    <tools-credential-generator />
     <bit-item>
       <button
         type="button"
@@ -22,19 +17,8 @@
     </bit-item>
   </ng-container>
   <ng-container bitDialogFooter>
-    <div class="modal-footer">
-      <button
-        type="button"
-        class="primary"
-        (click)="applyCredentials()"
-        appA11yTitle="{{ 'select' | i18n }}"
-        bitDialogClose
-      >
-        <i class="bwi bwi-lg bwi-fw bwi-check" aria-hidden="true"></i>
-      </button>
-      <button type="button" data-dismiss="modal" (click)="clearCredentials()" bitDialogClose>
-        {{ "cancel" | i18n }}
-      </button>
-    </div>
+    <button type="button" bitButton bitFormButton buttonType="secondary" bitDialogClose>
+      {{ "close" | i18n }}
+    </button>
   </ng-container>
 </bit-dialog>
