+++ apps/desktop/src/vault/app/vault/credential-generator.component.html
@@ -1,0 +1,40 @@
+<bit-dialog #dialog dialogSize="large" background="alt">
+  <span bitDialogTitle>{{ "generator" | i18n }}</span>
+  <ng-container bitDialogContent>
+    <ng-container *ngIf="data.type === 'username'">
+      <tools-username-generator (onGenerated)="onCredentialGenerated($event)" />
+    </ng-container>
+    <ng-container *ngIf="data.type === 'password'">
+      <tools-password-generator (onGenerated)="onCredentialGenerated($event)" />
+    </ng-container>
+    <bit-item>
+      <button
+        type="button"
+        bitLink
+        linkType="primary"
+        bit-item-content
+        aria-haspopup="true"
+        (click)="openHistoryDialog()"
+      >
+        {{ "generatorHistory" | i18n }}
+        <i slot="end" class="bwi bwi-angle-right" aria-hidden="true"></i>
+      </button>
+    </bit-item>
+  </ng-container>
+  <ng-container bitDialogFooter>
+    <div class="modal-footer">
+      <button
+        type="button"
+        class="primary"
+        (click)="applyCredentials()"
+        appA11yTitle="{{ 'select' | i18n }}"
+        bitDialogClose
+      >
+        <i class="bwi bwi-lg bwi-fw bwi-check" aria-hidden="true"></i>
+      </button>
+      <button type="button" data-dismiss="modal" (click)="clearCredentials()" bitDialogClose>
+        {{ "cancel" | i18n }}
+      </button>
+    </div>
+  </ng-container>
+</bit-dialog>
